+++ pubspec.yaml
@@ -66,6 +66,7 @@ dependencies:
   flutter_image_compress: ^2.4.0
   flutter_localizations:
     sdk: flutter
+  url_launcher: ^6.3.1
